remaking price graph for specific card view
