+++ preordain/static/preordain/src/app.postcss
@@ -60,6 +60,9 @@ main {
     button.preordain-button.delete {
         @apply enabled:hover:bg-red-600 transition-colors disabled:opacity-50
     }
+    a {
+        @apply hover:text-indigo-500 hover:underline transition-all
+    }
 }
 
 @layer utilities {
