+++ preordain/static/preordain/src/routes/Card.svelte
@@ -1,8 +1,9 @@
 <script lang="ts">
-    import { onMount } from "svelte";
     import { database } from "../util/fetch_data";
     import { connectURL } from "../assets/stores";
-    import { parseCurrency } from "../assets/stores";
+    import PriceGraph from "../lib/charts/PriceGraph.svelte";
+    import { parseOracleText, parseFormatLegality, parseCurrency } from "../util/dataFormatter";
+    import { link, push } from "svelte-spa-router";
     export let params = null
 
     async function getImage() {
@@ -10,47 +11,11 @@
         return URL.createObjectURL(await cardImg.blob())
     }
 
-    function parseOracleText(oracle: string) {
-        let respHTML = ""
-        let textLines = oracle.split("\n")
-        textLines.forEach(e => {
-            respHTML += `<li class="ml-8 leading-relaxed list-disc">${e}</li>`
-        });
-        return `
-            <div class="text-left font-semibold">Oracle:
-                <ul class="list-inside font-normal">
-                    ${respHTML}
-                </ul>
-            </div>
-            `
-    }
-
-    function parseFormatLegality(format: string) {
-        switch (format) {
-            case "legal":
-                return `<span class="bg-green-600/80   text-black border-2 border-black rounded-xl text-lg font-normal">
-                    Legal
-                </span>`
-            case "restricted":
-                return `<span class="bg-orange-500 text-white border-2 border-black rounded-xl text-lg font-normal">
-                    Restricted
-                </span>`
-            case "banned":
-                return `<span class="bg-red-500/80 text-white border-2 border-black rounded-xl text-lg font-normal">
-                    Banned
-                </span>`
-            default:
-                return `<span class="bg-gray-500   text-black border-2 border-black rounded-xl text-lg font-normal">
-                    Not Legal
-                </span>`
-        }
-    }
-
 
 </script>
 {#if params.cn && params.set}
     {#await database(`${$connectURL}/card/metadata/${params.set}/${params.cn}`) then data}
-    <div class="gap-4 flex flex-col">
+    <div class="gap-4 flex flex-col h-full">
 
         <!-- Card information -->
         <div class="component-theme w-full">
@@ -72,17 +37,18 @@
 
                     <div class="container mx-auto component-theme  px-4 py-1 leading-loose">
                         <div class="text-right"><span class="font-semibold">Mana Cost: </span>{data.data.mana_cost}</div>
-                        <!-- <br> -->
                         <div class="text-left"><span class="font-semibold">Type:</span> Legendary Creature - Human Soldier</div>
                         {@html parseOracleText(data.data.oracle_text)}
                         <div class="text-left"><span class="font-semibold">Artist: </span>{data.data.artist}</div>
                     </div>
 
-                    <div class="container mx-auto component-theme px-4 py-1 leading-loose">
+                    <div class="container mx-auto component-theme px-4 leading-loose py-3">
                         Legality
                         <div class="grid grid-cols-4 gap-2">
                             {#each Object.keys(data.data.legality) as format}
-                                    <span class="font-semibold">{format[0].toUpperCase()+format.slice(1)}</span>  <!-- Very hacky way of doing "Capitalize the first character"-->
+                                    <span class="font-semibold text-left px-4">
+                                        {format[0].toUpperCase()+format.slice(1)}
+                                    </span>  <!-- Very hacky way of doing "Capitalize the first character"-->
                                     {@html parseFormatLegality(data.data.legality[format])}
                             {/each}
                         </div>
@@ -94,12 +60,16 @@
             </div>
         </div>
 
-        <div class="flex flex-row gap-4">
-            <div class="grow component-theme">asd</div>
+        <div class="flex flex-row gap-4 grow">
+            <div class="h-full grow component-theme">
+                <PriceGraph set={params.set} id={params.cn} cardName={data.data.name}/>
+            </div>
+
+
 
             <!-- Other Variants -->
             <div class="component-theme w-1/4 text-center pb-4">
-                <span class="font-semibold ">Other Variants</span>
+                <span class="font-semibold">Other Variants</span>
 
                 <table class="table-auto w-full ">
 
@@ -119,8 +89,8 @@
                         {:then prices}
                             {#each prices.data as price}
                             <tr>
-                                <td class="text-left pl-4"><i class="ss ss-{price.set} ss-fw"/> {price.set_full}</td>
-                                <td class="text-right tabular-nums pr-4">{price.usd ? parseCurrency(price.usd, "usd") : `🌟 ${parseCurrency(price.usd_foil, "usd")}`}</td>
+                                <td class="text-left pl-4"><i class="ss ss-{price.set} ss-{price.rarity} ss-fw"/> <a href="/card/{price.set}/{price.id}" use:link> {price.set_full}</a></td>
+                                <td class="text-right tabular-nums pr-4">{price.usd ? parseCurrency(price.usd, "usd") : price.usd_foil ? `🌟 ${parseCurrency(price.usd_foil, "usd")}` : "N/A"}</td>
                             </tr>
                             {/each}
                         {/await}
