--- preordain/static/preordain/src/util/parseValues.ts
@@ -1,21 +1,0 @@
-// Check percentage
-export function parsePercentage(percent:string) {
-    if (percent === null) {
-        return ""
-    }
-    let change = parseFloat(percent)
-    if (Number.isNaN(change)) {
-        return "<p class='invisible'>empty</p>"
-    }
-    let classes = "text-white tabular-nums"
-    let emoji = "🤷‍♂️"
-    if (change > 0) {
-        emoji = "📈  "
-        classes = "text-emerald-500 tabular-nums"
-    } else if (change < 0) {
-        emoji = "📉"
-        classes = "text-rose-500 tabular-nums"
-    }
-    return `<span class=" tabular-nums">${emoji}</span>
-        <p class="${classes} text-right">${change}%</p>`
-}
